- Fix strndup implementation - Bump PORTREVISION
Reported by:	marcus
Patched by:	Max Brazhnikov

+++ devel/xdg-user-dirs/files/patch-xdg-user-dirs-update.c
@@ -5,7 +5,7 @@
  #include <langinfo.h>
  
 +#if defined(__FreeBSD__)
-+static char *strndup(const char *str, int len) {
++static char *strndup(const char *str, size_t len) {
 +	char *ret;
 +	
 +	if ((str == NULL || len < 0)) return(NULL);
